Remove extra testing target
